+++ ecosystem.config.js
@@ -2,7 +2,7 @@ module.exports = {
   apps: [
     {
       name: 'kreyolopal',
-      script: 'app/index.js',
+      script: 'server/index.js',
       instances: 'max',
       env: {
         NODE_ENV: 'development',
@@ -12,18 +12,4 @@ module.exports = {
       },
     },
   ],
-
-  deploy: {
-    production: {
-      user: 'SSH_USERNAME',
-      host: 'SSH_HOSTMACHINE',
-      ref: 'origin/master',
-      repo: 'GIT_REPOSITORY',
-      path: 'DESTINATION_PATH',
-      'pre-deploy-local': '',
-      'post-deploy':
-        'npm install && pm2 reload ecosystem.config.js --env production',
-      'pre-setup': '',
-    },
-  },
 }