+++ frontend/pages/LoginPage.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react'
-import LoginForm from '../components/LoginForm'
+import { Link } from 'react-router-dom'
+import LoginForm from '../components/Forms/LoginForm'
 import StandardPage from '../layouts/StandardPage'
 import {
   Box,
@@ -67,6 +68,10 @@ const LoginPage = () => {
           )}
 
           <LoginForm buttonLabel="Login" onSubmit={onSubmit} />
+          <hr />
+          <div className="has-text-centered">
+            <Link to="/resetpwd/request">An obliyé modpas an-mwen !</Link>
+          </div>
         </Box>
       </Section>
     </StandardPage>