+++ frontend/pages/SignupPage.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import LoginForm from '../components/LoginForm'
+import LoginForm from '../components/Forms/LoginForm'
 import StandardPage from '../layouts/StandardPage'
 import { useZakari } from '../components/ZakProvider'