+++ server/index.js
@@ -1,8 +1,5 @@
 const path = require('path')
 
-// ignore `.scss` imports
-require('ignore-styles')
-
 // transpile imports on the fly
 require('@babel/register')({
   configFile: path.resolve(__dirname, '../.babelrc'),