+++ server/routes/profile.js
@@ -3,6 +3,7 @@ const { body, validationResult } = require('express-validator')
 const passport = require('passport')
 const db = require('../database/models')
 const { protectedRoute } = require('../services/lib.auth')
+const authService = require('../services/authService')
 
 const { User } = db
 
@@ -69,6 +70,150 @@ const profile_route = ({ logger }) => {
           })
         },
         (reason) => {
+          logger.error(reason)
+          return res
+            .status(500)
+            .json({ status: 'error', error: 'Internal error' })
+        }
+      )
+    }
+  )
+
+  router.post(
+    '/updatepwd',
+    protectedRoute,
+    [
+      // password must be at least 6 chars long
+      body('currentPassword'),
+      body('newPassword').isLength({ min: 5 }),
+      body('verification').custom((value, { req }) => {
+        if (value !== req.body.newPassword) {
+          throw new Error('Password confirmation does not match password')
+        }
+        return true
+      }),
+    ],
+    async (req, res) => {
+      // Finds the validation errors in this request and wraps them in an object with handy functions
+      const errors = validationResult(req)
+      if (!errors.isEmpty()) {
+        return res.status(422).json({ status: 'error', error: errors.array() })
+      }
+
+      const { currentPassword, newPassword } = req.body
+
+      const user = await User.findByPk(req.user.id)
+      if (user === null) {
+        logger.error(`Cannot find user ${req.user.email} in database`)
+        return res.status(401).json({
+          status: 'error',
+          code: 401,
+          message: 'Unauthorized',
+          error: new Error('Unauthorized'),
+        })
+      }
+
+      const isValid = await authService.verifyPassword(
+        user.password,
+        currentPassword
+      )
+
+      if (!isValid) {
+        return res.status(401).json({
+          status: 'error',
+          code: 401,
+          message: 'Unauthorized',
+          error: new Error('Unauthorized'),
+        })
+      }
+
+      return authService.hashPassword(newPassword).then(
+        (hashedPassword) => {
+          user.password = hashedPassword
+          user.save({ fields: ['password'] }).then(
+            () => {
+              res.status(200).send({
+                status: 'success',
+                data: {},
+              })
+            },
+            (reason) => {
+              logger.error(reason)
+              return res
+                .status(500)
+                .json({ status: 'error', error: 'Internal error' })
+            }
+          )
+        },
+        (reason) => {
+          logger.error(`cannot hash password for ${req.user.email}`)
+          logger.error(reason)
+          return res
+            .status(500)
+            .json({ status: 'error', error: 'Internal error' })
+        }
+      )
+    }
+  )
+
+  router.post(
+    '/resetpwd',
+    [
+      // password must be at least 6 chars long
+      body('token')
+        .notEmpty()
+        .custom((value) => {
+          const isToken = /[A-Za-z0-9]{64}/
+          return value.length > 0 && isToken.test(value)
+        }),
+      body('newPassword').isLength({ min: 5 }),
+      body('verification').custom((value, { req }) => {
+        if (value !== req.body.newPassword) {
+          throw new Error('Password confirmation does not match password')
+        }
+        return true
+      }),
+    ],
+    async (req, res) => {
+      // Finds the validation errors in this request and wraps them in an object with handy functions
+      const errors = validationResult(req)
+      if (!errors.isEmpty()) {
+        return res.status(422).json({ status: 'error', error: errors.array() })
+      }
+
+      const { token, newPassword } = req.body
+
+      const user = await User.findOne({ where: { reset_pwd_token: token } })
+      if (user === null) {
+        logger.error(`Cannot find user in database`)
+        return res.status(401).json({
+          status: 'error',
+          code: 401,
+          message: 'Unauthorized',
+          error: new Error('Unauthorized'),
+        })
+      }
+
+      return authService.hashPassword(newPassword).then(
+        (hashedPassword) => {
+          user.password = hashedPassword
+          user.save({ fields: ['password'] }).then(
+            () => {
+              res.status(200).send({
+                status: 'success',
+                data: {},
+              })
+            },
+            (reason) => {
+              logger.error(reason)
+              return res
+                .status(500)
+                .json({ status: 'error', error: 'Internal error' })
+            }
+          )
+        },
+        (reason) => {
+          logger.error(`cannot hash password for ${user.email}`)
           logger.error(reason)
           return res
             .status(500)